requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.10
asyncpg==0.29.0
yoyo-migrations==9.0.0
setuptools
